Updated malua (rc1)
This is the first release candidate "build" from malua, so from now on I'll try not to change the code base that much.  The only thing I'm planning to do that's going to change the code base is to porting the framework to TypeScript at some point.

So here's what's changed:

Added widget creation helpers; (this is still going to be improved over time with more methods being added.)

Refactored most part of the widget code to utilize the widget creation helpers;

Simplified the HTML syntax when building the menu; (this is also because of the widget creation helpers.)

Made some minor changes in the `figure` and `button` widget css file; (position and size adjustments.)