Updated malua (rc3)
Implemented a new widget: ListBox.

Improved the method that malua uses to set the default widget value/state. (Now it's set before the widget initialization. I'm using an approach similar to React's life cycle.)

Added a fallback route for the setValue function to handle setting the selected index of ListBox and eventually ComboBoxes.