Updated malua (rc3.2)
Fixed a few "compatibility" issues related to the fit-content property and added missing title properties.

Improved GroupBox shadow root and fixed an issue where the GroupBox legend element would have a background applied to it when setting a shader for the GroupBox.

Improved ListBox init process.

Adjusted blur shaders to be less "blurry" (couldn't think of a better way to explain this)

Improved the setLabel function.